fixup: Allow RequestId to be treated as an integer (by legacy code)
... undoing dozens of arguably unnecessary changes introduced by the
previous branch commit.

An ideal code would probably treat the Request ID as an opaque key, and
that is the approach the previous commit used. However, there is so much
legacy code that treats this ID as a semi-opaque integer with a special
zero value that perhaps it is best to allow for such interpretation just
to reduce the footprint of these changes.