fixup: Removed the last known bare "unsigned int" RequestId::Index cases
Also added an out-of-scope TODO. While we were semi-force to change
AddMap() because it used a signed integer and a very odd/unnecessary
linear search algorithm, all other use cases can remain intact in this
branch. If we start changing them, we should probably hide the integer
nature of RequestId::Index from the map users, undoing the previous
branch commit.