fixup: Undo branch split of Mgr::Response sender's constructor
I am worried that some old code somewhere (not in master, I checked)
might call this constructor with a nil pointer -- something that the old
API did allow. It is possible that the correct approach here is to split
Mgr::Response into two classes rather than making Action optional, but I
am not sure, and such a split is outside this branch scope anyway.