Send Uds messages in the right order
IPC messages are scheduled for sending by Ipc::SendMessage() calls,
which spawns an UdsSender job for each message. On sending error,
the job retries the attempt after a while. As a result, the order of
sent messages can be different from the order of SendMessage()
calls, which the receiver does not expect.

This simple approach queues the messages before sending them, allowing
only a single UdsSender job for a process at a time.