Fixed stopPeerClock() timing, which was called too early
This method is called when all communication with peers completes.
There are many different paths when we need to call it, especially
on error path. To simplify this and protect ourselves from
forgetting calling this method I think we should do it once in
TunnelStateData destructor.

Also removed unnecessary checks for 'request' existence: the
most of other methods do not do this and we have checked already
in the constructor.