Build an Ssl::ServerBump object on all of the bumping cases
Currently squid does not build/use  Ssl::ServerBump object in the
case of client-first bumping mode. This is for two major reasons:
1) It does not need to store any info in client-first bumping mode
and 2) because this is how this code initially build and we never
fix it.
However there are info which it is better to be stored in an object
like the Ssl::ServerBump in all bumping cases. A such case is the
indirect client address of the parent CONNECT request.

----

Cherry-picked SQUID-616-Apply-Indirect-client-IP commit
c9205bfaa8cc39b0e7bd2a1545af785de2e18a1a