Store the received body length in HttpReply instead of MemObject
In the previous approach, this length was stored for virgin reply and
then (sometimes) overwritten by the adapted reply. It seems unreliable
because we need manually distinguish these two cases so that we would
not overwrite the adapted value with the virgin one.

Now this is up to each HttpReply object (virgin or adapted) to store
the length: StoreEntry just uses the value from the HttpReply it deals
with.

Also eliminated code duplication in StoreEntry::validLength().