Call FwdState::completed() regardless of the server connection
This (controversial) "Comm::IsConnOpen(serverConn)" condition was added
to FwdState::complete() at 792725ad. Later (since 7a0fb323), this method
started to initiate FwdState self-destruction (and the destructor, in
turn, calls FwdState::completed() unconditionally).

On one hand, these two calls of completed() (one with the condition and
one without it), following each other seem inconsistent - either we
should the apply same condition for both of them or we should not use it
at all. On the other hand, FwdState should do cleanup anyway (regardless
of the server connection), i.e., we should call completed()
unconditionally in both places. (we could even remove the first
completed() relying on the destructor only - but we want to
'complete' the entry ASAP and ref counting could take a time to
kick in).