Provide ALE with server-side error information
such as 'err_code', 'err_detail' and LogTags suffix for the '%Ss' code.
These fields remained empty if the client has started receiving the
response (i.e., the entry was not empty by that time).
On one hand, this behavior looks compliant with the documentation
describing 'err_code' as the "ID of an error response served by Squid".
On the other hand, Squid does not always follow this rule. For example,
it logs 'ERR_LIFETIME_EXP' without sending this code to the client.
Considering this, we decided that it is better to correct the
documentation and always add the important error information into
access.log. (TODO: adjust the documentation accordingly).

Also add a client-side specific 'error_detail' string, when it is empty
(for cases without errno). This should help distinguishing identical
client-side and server-side errors (such as timeouts).