Peek/stare mode: mimic server side TLS version to client-side connection
Also this patch removes an in practice NOP block of code which supposed
that it allow OpenSSL SSL_accept function to receive and check
client TLS hello message. This is never worked because before the
connections ClientBio object is in hold mode, thus no bytes are
transfered to OpenSSL internals.

It is important that the OpenSSL has not receive any bytes when the
TLS version will be set, in later step and after the connection with the
TLS server is established.