Prohibit meaningless response delay pools configurations
For example, a configuration where one of the individual-restore and
individual-maximum pair is zero and the other is nonzero will stall
client transactions regardless of the non-zero parameter value. The same
is true for the aggregate-restore/aggregate-maximum pair and, hence, for
the combination of all four parameters.

We should avoid such confusing/ambiguous configurations, forcing user to
specify either zero or non-zero values for all four parameters.