Remove the CertError object class and replace with Security::ErrorDetail
The two object classes duplicates the same error information.
Also this patch, adsd a broken_cert depth member to
Security::Erroretail object.

The ACLFilledCheckList object fixed to consider its sslErrors member as
a pointer to an externaly controled object, the ACLFilledChecklist
does not delete or removes any reference to the linked object any more.
Also the ACLFilledChecklist retrieves the CertErrors list from the
linked ConnStateData object, if exists and the ACLFilledChecklist::sslError
is not given. This new policy needs to be reviewed.