Quell "BUG: Unexpected state while connecting...", part 1
Do not show BUG messages and do not send error notifications when the
Security::PeerConnector user cancels its callback. That state is not
a bug.

Still show BUG messages on exceptions but add the job stop reason to the
message. This may help us confirm the actual reason behind those
messages. After that, we should convert the BUG message to a regular
debugging message because some of the exceptions are not BUGs (and those
that are should be announced when they happen, not in swanSong()).

Also removed wrong/stale XXX. If callback is nil, we have already called
disconnect(). Otherwise, we will it in bail(). Ether way, no additional
cleanup is required.