malformed_request_header_edit directive
This directive allows to adjust the received HTTP request headers that
Squid considers so malformed that they have to be rejected.  These
optional adjustments are performed once per request, as a last resort
measure. If no adjustments apply or if Squid still rejects the adjusted
header, the whole request is rejected.