Prefer 'const auto &' in ConfigParser's for-range loops
This adjustment will not optimize the existing loops, because Tokens
returns a newly created TokenIterator (as an rvalue) - so 'const auto'
and 'const auto &' are equivalent in those contexts. However, this
assumption relies on the present Tokens' implementation and  other
implementation (based, e.g., on caching the values in an array) would
surely benefit from this change.