Must clause about bumping step inside PeekingPeerConnector fails
Current squid code changes the bumping step to be step3 after the
PeekingPeerConnector started just before runs the ssl_bump access list.

Also while the PeekingPeerConnector intialised, inside the
PeekingPeerConnector::initialize method, there is Must clause to
check that when the PeekingPeerConnector started, the bumping step
is step1 or step2.

However this Must clause will fail if the first try to connect to
the remote server fails and squid try a next avaiable path. The
first PeekingPeerConnector object, called to initialize the
connection to the first available path will change bumping step
to be step3 and then, will abort the connection. Any new
PeekingPeerConnector object created for the next available
paths, will find the bumping step set to step3.