fixup: Simplified DebugMessage::allowed(cacheChannel)
We check the message level earlier. If the removed check was meant to
_compliment_ (i.e. be more strict in some cases than) the level check
above it, then it was not helping because _db_print_early_message()
caller already filters non-major messages out, and we should not try to
filter the same message for the same reason twice.

Also added more XXXs.