fixup: Restored level-1 "squid -k..." output
We need to show "Processing: http_port..." with -k parse and similar
important messages when running in one of the "-k..." modes.

Also stop pointless accumulation of early messages for future cache.log
dumping once we learn that there will be no cache.log to dump to.

Also do not call Debug::parseOptions("ALL,1"). It is the branch default.