Simplified debug channel code by converting DebugChannel enum to classes
Channel-specific state and logic are now confined to one class/object
rather than spread around various "if (ch == ...)" statements.

This conversion was made possible, in part, by removing all the ugly
variadic parameters inherited from C code. They were actually pretty
much unused -- we were passing strings!

This significant simplification is also a performance optimization
because it eliminates one copy: The formed message -- buf.str() -- is
now written to the logs without copying it into a BUFSZ buffer first. We
only use that buffer to form the message prefix.

This commit does not introduce std::string into debug.cc. The official
code is already using std::string implicitly because it is using
std::ostringstream. I think it would be possible to replace that with an
SBuf-based code, but that is outside this project scope.