hack: Signal excessive number of shared listening port requests
Admins can monitor cache.log for the new message and restart Squid:

    ERROR: Suspiciously many pending requests for listening ports:...

When Coordinator crashes and restarts, workers appear to get stuck with
an ever-growing queue of delayed worker-Coordinator requests for
listening ports. The causes of this bug are not yet known -- it is not
clear why the first pending (a.k.a. "active") request does not timeout
in this case.

This is a quick and temporary change because the number of configured
ports is hard-coded to 3.