fixup: Fixed (and simplified?) the safe-to-dump-saved condition
The (MaxErrLogLevel >= EarlyMessagesMaxLevel) condition could miss cases
where MaxErrLogLevel increased after some to-be-saved messages were
rejected by the errlog channel.

Also fixed "logged some but not all saved messages" example that should
have been written as "-d0".

XXX: The new ErrLogEnabled() assertion in writeAllSaved() fails in
"-k parse" tests, but that to-be-fixed bug is not this change fault.