fixup: Partially addressed DebugMessages::insert() XXXs
I do not like this solution because waivingLevelRestrictions counter
feels excessively complex/fragile (but cannot be easily simplified
because it needs to support two concurrent level restrictions override).

XXX: Dumping saved messages breaks rejectSavedMessages logic when stderr
logging of to-be-saved messages is prohibited by default because
shouldLog() first rejects these to-be-saved messages and then is forced
to accept them when they are dumped, blocking subsequent dumps for not
good reason.

Also reduced fprintf(stderr) code duplication.