fixup: Simplified by removing class DebugMessages
The class was essentially used to collect overflow-related statistics
but that statistics was difficult to interpret correctly, especially for
admins, and will probably never be shown (because we do not expect
overflows). It also needed more work to become more usable. It just was
not worth the complexities it has created.

Also use std::deque instead of std::vector for early message storage. We
do not access that storage at random offsets, so it is probably best to
optimize away memory copying (if the number of early messages and STL
implementation defaults force more than one memory allocation).