fixup: Do not drop the message that overflowed early message storage
Our new LogMessage()-order enforcement works. It implies that we should
not save or log messages out of order because doing so will effectively
block the messages we cheated out of their chance to be saved/logged
from being logged. Fortunately, only the lines between LogMessage() and
maybeLog()'s lastLoggedRecordNumber assignment (or saveMessage()'s
emplace_back) need to be careful about this.