Removed ConfigParser::Undo() mechanism
The 'undo' functionality was created primarily for re-parsing '-i' and
'+i' flags if these flags were specified in the general ACL flags
section. This approach has a flaw, because when an AclData started
parsing ACL parameters (calling ConfigParser::NextToken()), it was
unable to tell whether the first returned token is an 'undone' flag
(e.g., -i) or an ACL argument with '-i' value.

Now all ACL flags (including -i and +i) are parsed uniformly. After
that, the parsed ACL data specific flags (-i and +i) are passed as
default/initial values for ACL parameters.