Update http_access ACL pointers after smooth_reconfiguration
This commit adds a general mechanism for updating a given ACL-storing
directive and applies that mechanism to http_access. The mechanism
attempts to prevent reconfiguration crashes/inconsistencies by doing a
reconfiguration rehearsal or "dry run" -- if dry run finds a directive
with a disappeared ACL, reconfiguration is aborted before taking effect.

ACL updates are based on admin-given ACL names and ACL::FindByName().
XXX: External ACLs uses FindByName() for async operations. That probably
means that a single rule involving external ACLs may start using one set
of ACLs and finish with another. This hidden and silent inconsistency is
a bug that should be fixed.