Revised the mechanism for updating a given ACL-storing directive
... added two branch commits ago (commit 030cbc0) to address its XXX.

We cannot update the ACL tree pointed to by Config.accessList.foo
because that tree may be in-use by Checklists and other objects. We must
create a brand new tree.

Since the new code has to create new trees, it is a lot more difficult
(and a bit more risky) to propagate dryRun into the tree creation code.
It is easy to write a _separate_ set of methods that will check the
feasibility of updating a given tree, but then the tree creation code
can easily get out of sync with the checking code. The currently
implemented approach minimizes that out-of-sync possibility at the
expense of using a few more CPU cycles (and using extra RAM a bit
longer): It creates each ACL tree twice. During dryRun, created trees
are discarded.

When creating new ACL trees, we do not create new _named_ ACLs. Those
are created, earlier when parsing "acl aclname typename" lines.

Also supplied new ACL tree nodes with names and squid.conf locations.