Do not tie StoreEntry completion timing to completion outcome; other
Previous branch changes called StoreEntry::completeSuccessfully() as
soon as we knew that the whole entry was stored. That is fine in
principle, but StoreEntry::complete*() has many side effects and the
official code checks a lot of StoreEntry/reply properties on the way to
that StoreEntry::complete*() call. I am worried that, by changing
StoreEntry/reply state much earlier than the official code expects, we
will alter some of the code decisions that our test cases do not tickle.

For example, FwdState::reforward() tests a lot of StoreEntry, reply,
request, and other transaction properties. Some of them may change after
FwdState::bodyReceivedSuccessfully/markStoredReplyAsWhole() and before
the FwdState::complete() call. Starting to assume/rely on reforward()
reusability is too dangerous IMO because the method is complex and was
not maintained as a reusable call.

This commit undoes the changes in StoreEntry::complete*() _timing_ by
simply remembering the "we stored the whole entry" information until
it is time to call StoreEntry::complete*().

Renamed a few methods to underscore the difference between "receiving"
(from a peer) and "storing" (into Store). I even suspected that some
methods were called in the wrong places, for the wrong reasons, and
started to fix those imaginary bugs until I realized that the methods
were not about receiving or bodies!. They are about storing replies.

I also realized that StoreEntry completeness decisions are based not so
much on bytes we "received" but on the bytes we "parsed". For example,
we could receive the entire response but unable to parse its last-chunk
due to unsupported extensions/etc and, hence, unable to store the entire
response. The difference is important because our receiving (i.e. socket
reading) code is quite separated from our parsing code. If we needed to
make "received" decisions, it would be tempting to make those as we read
the reply bytes, but that would be impossible in many cases.

Our parsing code is also separated from our storing code. An earlier
version of this commit remembered "parsed the whole reply" events
instead, but I replaced that complication with a
Client::markParsedVirginReplyAsWhole() comment because the current
addVirginReplyBody() method cannot tell whether it was given everything
we parsed. Thus, it would have to assume almost as much as the
being-committed (and simpler) code assumes.

Finally, I think it was very important to emphasize that
virginBodyReceivedSuccessfully() only works for virgin responses.
Adapted responses have their own markStoredReplyAsWhole() calling logic.

XXX: The changes exposed a possible Gopher support bug in the branch.

XXX: I still do not like the method actual spelling/wording. Hopefully,
we will find a better name that still addresses the above concerns.

We must not mark aborted adapted responses as whole. The API already
tells us about such aborts by calling handleAdaptedBodyProducerAborted()
instead of handleAdaptedBodyProductionEnded().

AFAICT, FTP code could call that method sooner, so I moved those lines.
The timing of the call may be important if something bad happens after
the call (e.g., we fail to send a confirmation to the FTP agent after
parsing the entire response).

State, log, and preserve justification for the markStoredReplyAsWhole()
and similar decisions to help with triage. Otherwise, they would all be
cramped into a one "truncated entry" reason in completeUnsuccessfully().