fixup: Addressed ENTRY_ABORTED branch XXX
We could modify all markStoredReplyAsWhole() callers to check
ENTRY_ABORTED, but that seems like a lot of code duplication and a good
opportunity to forget to check. Instead, we are essentially interpreting
"wrote to store" or "stored" in these new method names as "gave Store
the data"; it is then up to Store whether to actually store anything or
not, of course.