fixup: Restored STORE_PENDING assertion in FwdState::reforward()
I am not sure the code needs that assertion. We already have a
ENTRY_FWD_HDR_WAIT check that tries to test the same thing. I suspect
replacing that assertion with an `if` statement is a good idea. None of
that is relevant to this branch changes anymore because the branch calls
reforward() when the official code does and with the same entry state.