Reworked to simplify
We are paying with one extra memory allocation for that simplicity, but
I think it is worth the cost, especially since this only happens on the
error handling path. I am sure resumeNegotiation() can be refactored
further to remove that extra memory allocation, but it is probably best
to leave that refactoring out of this bug fix.