Do more to log early messages and assertions
The Debug module has to save early messages to avoid their loss. Doing
so means not logging some messages immediately. When Squid is asserting,
it cannot postpone logging the assertion (for much longer). The branch
code relied on the Debug module to dump all the early messages and the
assertion message. That was too dangerous because it is the Debug module
code dedicated to saving early messages that could assert!

Instead, we should dump early messages (if any) and enable emergency
stderr use (if cache.log has not been opened) to prepare for immediate
and simplified logging of the assertion message.

This change may also improve support for stack dumping in death()
because we now stop using stderr before that low-level hack assumes
control over that channel.

The SwanSong name was no longer appropriate because, in most use cases,
we do expect at least one more debugs() call after that method call.