Do "less" to log early messages and assertions
This commit revises earlier branch commit 4773a79 "Do more to log early
messages and assertions". The primary principles behind that older
commit were correct, but it lacked understanding that an assertion
triggered from debugs() requires either correct reentrant code or
protection from reentrancy (that works even for such asserts). That
earlier code had neither. Now that we have reentrancy protection we
should use it, even for asserts.

The old commit comment regarding stack dumping in death() was not wrong,
but we should fix that code by (updating and) moving it to debug.cc
rather than placing Debug::PrepareToDie() in an unusual spot because of
it.