fixup: Delayed Debug::Context->DebugMessageHeader conversion
... to reduce dangerous DebugMessageHeader dependency on the global
Current Context and to simplify.

This improvement also clarifies the relationship between Debug::Context
and DebugMessageHeader. The former is an "active" or "live" debugs()
message. The latter is meta-information of a Finish()ed or "frozen"
debugs() message; the frozen message body is stored separately.

It is tempting to merge Debug::Context, DebugMessageHeader, and the
frozen message body into one class, but doing so is tricky because we
need to copy [frozen] debugs() messages but Context::buf (ostringstream)
cannot be easily copied. We can work around that limitation

a) by refcounting Context objects (each stream stores the same message
   today) but that idea conflicts with a fairly important Debug::Start()
   optimization that avoids pretty expensive creation of new context
   objects for the vast majority of debugs() messages. We would need to
   implement custom Context pooling to mimic that optimization.

b) by creating a new Content::buf and filling it with the other
   Context::buf contents. TODO?