fixup: Read.cc:69: assert "fd_table[conn->fd].halfClosedReader != NULL"
... when splicing during step3 (peek all / splice all).

This "two active-reads" assertion was triggered because branch commit
fc5b631 effectively allowed FwdState::completed() to call
ConnStateData::httpsPeeked() after establishing a TCP tunnel with the
server -- see the answer.tunneled case in FwdState::connectedToPeer().
When both TunnelStateData and ConnStateData start reading from the same
connection, we (thankfully) assert.

I do not think this bug is related to the later branch commit a4c0c19.

I am not sure whether we should call entry->completeTruncated() instead
of adding/setting flags.abandonEntry. It feels wrong to "complete" an
entry that received no bytes, even if it is going to be marked as
truncated -- we did not really truncate anything. The tunneling code
does not use Store at all. I am worried about StoreEntry::complete()
calling invokeHandlers() and perhaps activating ConnStateData that way.
On the other hand, ConnStateData was probably not activated in the
official code that was calling StoreEntry::complete(). It would be good
to figure out why.

I am also not sure what prevented the official code from reforward()ing
the established SslBump tunnel (in some cases). I set flags.dont_retry
to prevent that, but we should understand why it was not needed before.