fixup: Documented GopherStateData::len
We had to figure out what that data member is in order to decide when
(not) to call markStoredReplyAsWhole(). Since Squid does not recognize
Gopher's reply-ending bare period (".") signal[^1], an incomplete line
is our only indication of an incomplete response.

[^1]: In Squid defense, according to a citation-free Wikipedia claim,
many servers do not send bare period. Look for "full-stop" at
https://en.wikipedia.org/wiki/Gopher_(protocol)