Use a new PeekingPeerConnector::currentBumpingMode member
... to store the current buping mode for use by PeekingPeerConnector
code. Currently in most cases the ConnStateData::sslBumpMode is used,
but this is can be buggy: the PeekingPeerConnector updates the
ConnStateData::sslBumpMode while checks ssl_bump access list.
If the PeekingPeerConnector fails and squid retry next IP address
starting a new PeekingPeerConnector job, the ConnStateData::sslBumpMode
will have wrong value.

This patch uses the new PeekingPeerConnector::currentBumpingMode
member where it stores the step2 bumping decision which is the
real value we are actually interested for.