Fix certificate and certs chain loading for https_port
Now while read certificates from tls-cert file builds the certiicates
chain with the correct order even if certificates are liste unordered
in this file. Self-signed certificates (eg root certificates),
certificates which are not issuers of an other certificate in chain,
or duplicated certificates are droped from the chain.

The tls-key file now may contain more than one keys, squid iterates
on all keys listed in this file to find the correct one.