fixup: Avoid "found KEY_PRIVATE" WARNING after branch commit f023e98
... as modified by follow-up commits.

The underlying logic of the !cacheWriter() check in the branch was "if
we are not a cache writer, we should not be swapping out". That logic is
correct, in principle, but it is too broad for this specific "if ...
then yield to publisher" check. For example, when dealing with a private
StoreEntry, we _are_ the writer/publisher (so we cannot yield the
swapout decision to ourselves), but cacheWriter() correctly returns
false because we write to Store, not cache (and we "publish" for a
single client).

Unlike !cacheWriter(), transientReader() is false for private entries so
we do not (incorrectly) yield the swapout decision to ourselves.
Instead, we mark the entry for release() in the checkCachable() clause
further below. That marking avoids the "found KEY_PRIVATE" WARNING.

Exposed by SMP Daft tests that involve non-aggregating cache manager
queries such as mgr:openfd_objects. Mgr::StoreToCommWriter::swanSong()
was unlocking a private StoreEntry that was not marked for release.

It is debatable whether the warning itsef is correct: On one hand,
private entry writers should not be forced to release() their creations;
Store could silently release() a private unlocked entry. On the other
hand, some private entries become public later; attracting attention to
cases where the code forgot to make that switch/decision may be useful.