Fix ServerOptions::{loadClientCaFile, updateContextClientCa} meethods
The commit "Fix ServerOptions::createStaticServerContext method" broke
the clientca https_port option. This commit moved the
ServerOptions::updateContextConfig method in a step before the
server certificates are loaded, to allow users use ciphers or TLS
options which affect certificates loading/checking.
Unfortunatelly this move has as result the updateContextClientCa
to be called before the loadClientCaFile is called.

This patch calls loadClientCaFile method from updateContextClientCa
to avoid similar bugs in the future.