fixup: Synced CertGadgets-related changes with yadij/gnutls-chainload
... as of that branch commit 350dc38.

The only non-trivial current difference I know about is that this branch
was checking for non-nil X509_get_subject_name() return. OpenSSL
documentation does not disclose whether that check is needed (AFAICT), I
could not quickly understand whether it is needed based on OpenSSL code,
and Squid code is very inconsistent in checking this return value. It
may be a good idea to check it, but that difference alone is not worth
keeping.

This code will change. I simplified tracking its evolution by copying
these three new/changed gnutls-chainload files "as is".

This commit also removes temporary Security::CertIsIssuedBy() debugging.