fixup: Removed ACL::lineOptions()
This used-once method was dangerous because it returned Options with
value lifetimes limited to the parseFlags() call. Its name and
documentation did not emphasize the temporary nature of those values.
Inlining that method inside parseFlags() reduces associated dangers.

Also made options() private, like currentLineOptions(). This improves
cohesion between those two "paired" methods. Perhaps more importantly,
it reduces the danger of [external code] calling options() and thinking
that it is getting _all_ options. Ideally, the method should be renamed
to globalOptions() but that will probably introduce a lot of PR noise.