Fixed configureDefault() name and clarified its meaning
The configureDefault() name was known to be difficult to comprehend. It
was trying to both capture the opposite meaning of "with [value]" and
accommodate (future/potential) "--disable-foo" use cases. After the
branch added Option::disable() and clarified that configureDefault() is
limited to enableName use cases, we can use an intuitive enable() name.

TODO: We should probably rename configureWith() to enableWith(), but
that method name does not cause a lot of comprehension problems (AFAIK).