The selected ciphers must affect certificates loading and use
The selected ciphers in an https port play role in the provided
security policy and the configured certificates should accepted
or rejected depending on whether or not comply to this policy.

An example is the fake @SECLEVEL cipher used by OpenSSL toolkit.

This patch fixes the ServerOptions::createStaticServerContext
and Ssl::createSSLContext to configure TLS context ciphers
before attach certificates to this context.

This is a Measurement Factory project