Correctly implement RFC 5077 Session Tickets extension for Squid
According Session Tickets TLS extension the client is responsible
to store last session parameters and send them to the TLS server
when negotiates new TLS connections. The session parameters are
encrypted using a secret key provided by the TLS server.
The TLS server needs to maintain and store the secret key to be
able to retrieve session parameters send by client.

This patch add a shared mem store to Squid for generated secret keys.
A new secret key generated every 3 hours. The generated keys are valid
for about 6 hours. The secret keys are generated by the coordinator
process.

This is a Measurement Factory project