Do not respond to the same CONNECT twice
TunnelStateData::clientExpectsConnectResponse() was using the
ServerBump::at(stepA, stepB) method to decide whether the SslBump has
progressed beyond step1 and, hence, has already responded to a real
CONNECT (if any). The corresponding `step == step2 or step == step3`
logic stopped working when we started setting the current ServerBump
step to bumpEnd at the end of SslBump processing. `step >= step1` would
have continued to work, but we should not assume the step values are
ordered. I am not sure what previous (branch?) change broke this.

Instead of trying to enumerate all possible cases where we have already
responded to the CONNECT request, we should simply remember that the
CONNECT request has been responded to. ClientHttpRequest is the right
place to keep that information (because HttpRequest containing CONNECT
is used in unrelated Squid-peer and adaptation communications).

The only caveat here is that we should distinguish fake CONNECTS from
the real ones because only the latter imply that the client expects a
response. There is only one such fake request creator, and new ones are
rather unlikely so using a defaulted constructor parameter is probably
safe enough.

Also completed a CONNECT generation de-duplication TODO.

Also fixed the "If we are forcing a tunnel after receiving a client
CONNECT, then we have already responded to that CONNECT" comment: We
never force a tunnel after receiving a real client CONNECT. Once we can
guarantee ClientHttpRequest lifetime exceeds TunnelStateData lifetime,
we can remove this and the other special conditions, removing the entire
TunnelStateData::clientExpectsConnectResponse() response method in favor
of its reliable ClientHttpRequest cousin.